Validateurs API

+++ node/queries/equipments.js
@@ -6,28 +6,54 @@ const check = require('../validators.js');
 // add query functions
 function getAllEquipments(req, res, next) {
   pool.query('select * from equipments',(err,rows) =>  {
-    if (err) throw err;
+    if (err) return errors(res,err);
     return res.send(rows.rows);
   })
 }
 
 function addEquipments(req, res, next) {
+  let body = check.checkForm(res,[check.hasProperties(["name"],req.body)])
+  if (body !== true) {
+    return body;
+  }
+
+  let verif = check.checkForm(res,[check.limitedText(req.body.name,40)])
+  if (verif !== true) {
+    return verif;
+  }
+
   pool.query('insert into equipments (name) values ($1)',[req.body.name],(err,rows) =>  {
-    if (err) throw err;
+    if (err) return errors(res,err);
     return res.send({data : true});
   })
 }
 
 function deleteEquipments(req, res, next) {
+  let verif = check.checkForm(res,[check.arrayOfValidFk(req.body)])
+  if (verif !== true) {
+    return verif;
+  }
+
   pool.query(format('delete from equipments where id in (%L)',req.body),(err,rows) =>  {
-    if (err) throw err;
+    if (err) return errors(res,err);
     return res.send({data : true});
   })
 }
 
 function updateEquipments(req, res, next) {
+  let body = check.checkForm(res,[check.hasProperties(["name","id"],req.body)])
+  if (body !== true) {
+    return body;
+  }
+
+  let verif = check.checkForm(res,[check.limitedText(req.body.name,40),check.validFk(req.body.id)])
+  if (verif !== true) {
+    return verif;
+  }
+
+
   pool.query('update equipments set name = $1 where id = $2',[req.body.name,req.body.id],(err,rows) =>  {
-    if (err) throw err;
+    if (err) return errors(res,err);
     return res.send({data : true});
   })
 }
