+++ node/queries/tasks.js
@@ -1,5 +1,7 @@
 var pool = require('../db.js')
 const format = require("pg-format");
+const errors = require('../errors.js');
+const check = require('../validators.js');
 
 // add query functions
 function getAllTasks(req, res, next) {
@@ -10,6 +12,18 @@ function getAllTasks(req, res, next) {
 }
 
 function addTasks(req, res, next) {
+  let body = check.checkForm(res,[check.hasProperties(["name"],req.body)])
+  if (body !== true) {
+    return body;
+  }
+
+  let verif = check.checkForm(res,[
+    check.limitedText(req.body.name,40)
+  ])
+  if (verif !== true) {
+    return verif;
+  }
+
   pool.query('insert into tasks (name) values ($1)',[req.body.name],(err,rows) =>  {
     if (err) throw err;
     return res.send({data : true});
@@ -17,6 +31,13 @@ function addTasks(req, res, next) {
 }
 
 function deleteTasks(req, res, next) {
+  let verif = check.checkForm(res,[
+    check.arrayOfValidFk(req.body)
+  ])
+  if (verif !== true) {
+    return verif;
+  }
+
   pool.query(format('delete from tasks where id in (%L)',req.body),(err,rows) =>  {
     if (err) throw err;
     return res.send({data : true});
@@ -24,6 +45,19 @@ function deleteTasks(req, res, next) {
 }
 
 function updateTasks(req, res, next) {
+  let body = check.checkForm(res,[check.hasProperties(["name","id"],req.body)])
+  if (body !== true) {
+    return body;
+  }
+
+  let verif = check.checkForm(res,[
+    check.validFk(id),
+    check.limitedText(req.body.name,40)
+  ])
+  if (verif !== true) {
+    return verif;
+  }
+
   pool.query('update tasks set name = $1 where id = $2',[req.body.name,req.body.id],(err,rows) =>  {
     if (err) throw err;
     return res.send({data : true});
